combine facades and move to domain
