+++ domain/src/main/java/com/github/ryjen/jokeapp/domain/arch/module/UseCaseModule.kt
@@ -1,5 +1,6 @@
 package com.github.ryjen.jokeapp.domain.arch.module
 
+import com.github.ryjen.jokeapp.domain.facades.JokeFacade
 import com.github.ryjen.jokeapp.domain.usecase.*
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
@@ -17,4 +18,6 @@ internal val useCaseModule = module {
     factory { AddFavoriteJoke(get()) }
 
     factory { RemoveFavoriteJoke(get()) }
+
+    factory { JokeFacade(get(), get(), get(), get()) }
 }
