+++ ui/src/main/java/com/github/ryjen/jokeapp/ui/arch/module/ViewModelModule.kt
@@ -1,20 +1,11 @@
 package com.github.ryjen.jokeapp.ui.arch.module
 
-import com.github.ryjen.jokeapp.ui.jokes.favourites.FavoriteJokesFacade
 import com.github.ryjen.jokeapp.ui.jokes.favourites.FavoritesViewModel
-import com.github.ryjen.jokeapp.ui.jokes.random.RandomJokeFacade
 import com.github.ryjen.jokeapp.ui.jokes.random.RandomJokeViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 internal val viewModelModule = module {
-
-    factory {
-        FavoriteJokesFacade(get(), get())
-    }
-    factory {
-        RandomJokeFacade(get(), get(), get())
-    }
 
     viewModel {
         FavoritesViewModel(get())
