+++ ui/src/main/java/com/github/ryjen/jokeapp/ui/jokes/favourites/ViewModel.kt
@@ -2,10 +2,11 @@ package com.github.ryjen.jokeapp.ui.jokes.favourites
 
 import androidx.lifecycle.viewModelScope
 import com.github.ryjen.jokeapp.domain.arch.redux.ReduxStore
+import com.github.ryjen.jokeapp.domain.facades.JokeFacade
 import com.github.ryjen.jokeapp.ui.arch.redux.ReduxViewModel
 
 class FavoritesViewModel(
-    facade: FavoriteJokesFacade
+    facade: JokeFacade
 ) : ReduxViewModel<FavoritesState, FavoritesViewState, FavoritesAction>() {
 
     override val store =
